django/requirements.txt
Django==1.8.5
git+ssh://git@github.com/bartekbrak/gwaith.git#egg=gwaith
djangorestframework==3.2.4
